style: cargo fmt + fix 5 clippy warnings
Clippy fixes:
- 4× clone-on-copy on AccountPosition (which became Copy after the
  Raw/Typed split): liquidation.rs, repay.rs, supply.rs, withdraw.rs.
  Replaced .clone() with *deref since the type is Copy.
- 1× unnecessary i128 -> i128 cast in fuzz_ttl_keepalive.rs.

Plus cargo fmt across the workspace — touches mostly certora
verification specs, view modules, and a few formatting drifts in
common/contracts. No behavior changes; cargo fmt --check passes,
cargo clippy --all-targets -- -D warnings passes, 784 tests pass.

+++ verification/test-harness/tests/fuzz_ttl_keepalive.rs
@@ -355,7 +355,7 @@ proptest! {
             let bps = partial_bps.get(i).copied().unwrap_or(5000);
             let cur = t.supply_balance_raw(user, asset);
             if cur <= 1 { break; }
-            let amt = (cur as i128 * bps as i128) / 10_000;
+            let amt = (cur * bps as i128) / 10_000;
             if amt > 0 && amt < cur {
                 let _ = t.try_withdraw(user, asset, amt as f64 / 10_f64.powi(7));
             }
